# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface LanguageBreakdown {
  language: string;
  files: number;
  percentage: number;
}

export interface IntelligenceOverview {
  projectName: string;
  lastScanAt: string;
  totalFiles: number;
  totalModules: number;
  totalDependencies: number;
  healthScore: number;
  riskCount: number;
  languageBreakdown: LanguageBreakdown[];
}

export type DependencyNodeRisk = typeof DependencyNodeRisk[keyof typeof DependencyNodeRisk];


export const DependencyNodeRisk = {
  high: 'high',
  medium: 'medium',
  low: 'low',
} as const;

export interface DependencyNode {
  id: string;
  label: string;
  path: string;
  directory: string;
  language: string;
  lines: number;
  imports: number;
  importedBy: number;
  risk: DependencyNodeRisk;
}

export type DependencyEdgeKind = typeof DependencyEdgeKind[keyof typeof DependencyEdgeKind];


export const DependencyEdgeKind = {
  import: 'import',
  reference: 'reference',
  generated: 'generated',
} as const;

export interface DependencyEdge {
  source: string;
  target: string;
  kind: DependencyEdgeKind;
}

export interface DependencyGraph {
  nodes: DependencyNode[];
  edges: DependencyEdge[];
  groups: string[];
}

export type CodebaseFileRisk = typeof CodebaseFileRisk[keyof typeof CodebaseFileRisk];


export const CodebaseFileRisk = {
  high: 'high',
  medium: 'medium',
  low: 'low',
} as const;

export interface CodebaseFile {
  path: string;
  name: string;
  directory: string;
  language: string;
  lines: number;
  complexity: number;
  risk: CodebaseFileRisk;
  riskLabel: string;
  summary: string;
}

export type FileDetail = CodebaseFile & {
  imports: string[];
  importedBy: string[];
  symbols: string[];
  excerpt: string;
};

export type RiskFindingSeverity = typeof RiskFindingSeverity[keyof typeof RiskFindingSeverity];


export const RiskFindingSeverity = {
  high: 'high',
  medium: 'medium',
  low: 'low',
} as const;

export interface RiskFinding {
  id: string;
  severity: RiskFindingSeverity;
  title: string;
  detail: string;
  file: string;
  impact: string;
  relatedFiles: string[];
}

export type ActivityItemType = typeof ActivityItemType[keyof typeof ActivityItemType];


export const ActivityItemType = {
  scan: 'scan',
  finding: 'finding',
  dependency: 'dependency',
  system: 'system',
} as const;

export interface ActivityItem {
  id: string;
  title: string;
  detail: string;
  timestamp: string;
  type: ActivityItemType;
}

export type ScanInputScope = typeof ScanInputScope[keyof typeof ScanInputScope];


export const ScanInputScope = {
  full: 'full',
  dependencies: 'dependencies',
  risks: 'risks',
} as const;

export interface ScanInput {
  scope?: ScanInputScope;
}

export type ScanResultStatus = typeof ScanResultStatus[keyof typeof ScanResultStatus];


export const ScanResultStatus = {
  accepted: 'accepted',
} as const;

export interface ScanResult {
  scanId: string;
  acceptedAt: string;
  scope: string;
  status: ScanResultStatus;
}

export type GetDependencyGraphParams = {
groupBy?: GetDependencyGraphGroupBy;
};

export type GetDependencyGraphGroupBy = typeof GetDependencyGraphGroupBy[keyof typeof GetDependencyGraphGroupBy];


export const GetDependencyGraphGroupBy = {
  directory: 'directory',
  layer: 'layer',
} as const;

export type ListCodebaseFilesParams = {
search?: string;
language?: string;
risk?: ListCodebaseFilesRisk;
/**
 * @minimum 1
 * @maximum 100
 */
limit?: number;
};

export type ListCodebaseFilesRisk = typeof ListCodebaseFilesRisk[keyof typeof ListCodebaseFilesRisk];


export const ListCodebaseFilesRisk = {
  all: 'all',
  high: 'high',
  medium: 'medium',
  low: 'low',
} as const;