# lib/api-zod/src/generated/types/activityItem.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ActivityItemType } from './activityItemType';

export interface ActivityItem {
  id: string;
  title: string;
  detail: string;
  timestamp: Date;
  type: ActivityItemType;
}