# lib/api-zod/src/generated/types/codebaseFile.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CodebaseFileRisk } from './codebaseFileRisk';

export interface CodebaseFile {
  path: string;
  name: string;
  directory: string;
  language: string;
  lines: number;
  complexity: number;
  risk: CodebaseFileRisk;
  riskLabel: string;
  summary: string;
}