# lib/api-zod/src/generated/types/dependencyGraph.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DependencyEdge } from './dependencyEdge';
import type { DependencyNode } from './dependencyNode';

export interface DependencyGraph {
  nodes: DependencyNode[];
  edges: DependencyEdge[];
  groups: string[];
}