# lib/api-zod/src/generated/types/fileDetail.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CodebaseFile } from './codebaseFile';

export type FileDetail = CodebaseFile & {
  imports: string[];
  importedBy: string[];
  symbols: string[];
  excerpt: string;
};