# lib/api-zod/src/generated/types/intelligenceOverview.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { LanguageBreakdown } from './languageBreakdown';

export interface IntelligenceOverview {
  projectName: string;
  lastScanAt: Date;
  totalFiles: number;
  totalModules: number;
  totalDependencies: number;
  healthScore: number;
  riskCount: number;
  languageBreakdown: LanguageBreakdown[];
}