# lib/api-zod/src/generated/types/scanInput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ScanInputScope } from './scanInputScope';

export interface ScanInput {
  scope?: ScanInputScope;
}