# lib/api-zod/src/generated/types/scanResult.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ScanResultStatus } from './scanResultStatus';

export interface ScanResult {
  scanId: string;
  acceptedAt: Date;
  scope: string;
  status: ScanResultStatus;
}